feat(produtos): implementar esquema de criação de produtos com o padrão facade

+++ EncantoPersonalizados/src/main/java/school/sptech/EncantoPersonalizados/repository/ItemProdutoRepository.java
@@ -6,6 +6,6 @@ import school.sptech.EncantoPersonalizados.entities.ItemProduto;
 import java.util.List;
 import java.util.Optional;
 
-public interface ItemProdutoRepository extends JpaRepository<ItemProduto, Long> {
+public interface ItemProdutoRepository extends JpaRepository<ItemProduto, Integer> {
     Optional<List<ItemProduto>> findByPrecoVendaLessThan(Double preco);
 }
